+++ Sources/SwiftiomaticKit/Configuration/Configuration.swift
@@ -285,7 +285,7 @@ package struct Configuration: Sendable, Equatable {
 
 /// A version number that can be specified in the configuration file, which allows us to change the
 /// format in the future if desired and still support older files.
-private let highestSupportedConfigurationVersion = 8
+package let highestSupportedConfigurationVersion = 8
 
 package extension Configuration {
     /// The URL of the JSON schema hosted on GitHub. Embedded as `$schema` by `encode(to:)` .